+++ ElvUI_LuckyoneUI/Core/Layouts.lua
@@ -596,7 +596,7 @@ function Private:Layout_TheWarWithin(layout)
 	E.db.tooltip.healthBar.fontOutline = Private.Outline
 	E.db.tooltip.healthBar.fontSize = (scaled and 9) or 10
 	E.db.tooltip.healthBar.height = 6
-	E.db.tooltip.healthBar.statusPosition = 'TOP'
+	E.db.tooltip.healthBar.statusPosition = 'DISABLED'
 	E.db.tooltip.inspectDataEnable = false
 	E.db.tooltip.itemCount.bags = false
 	E.db.tooltip.itemCount.bank = false
@@ -664,9 +664,39 @@ function Private:Layout_TheWarWithin(layout)
 
 	-- Shared Arena
 	E.db.unitframe.units.arena.customTexts = {}
-	E.db.unitframe.units.arena.customTexts.Luckyone_HP = { attachTextTo = 'Health', enable = true, font = Private.Font, fontOutline = Private.Outline, justifyH = 'LEFT', size = 12, text_format = '[luckyone:health:percent] • [health:current:shortvalue]', xOffset = 3, yOffset = 0 }
-	E.db.unitframe.units.arena.customTexts.Luckyone_Name = { attachTextTo = 'Health', enable = true, font = Private.Font, fontOutline = Private.Outline, justifyH = 'RIGHT', size = 12, text_format = '[luckyone:name:last-classcolor]', xOffset = -3, yOffset = 0 }
-	E.db.unitframe.units.arena.customTexts.Luckyone_Power = { attachTextTo = 'Power', enable = true, font = Private.Font, fontOutline = Private.Outline, justifyH = 'CENTER', size = 11, text_format = '[luckyone:power:percent-color]', xOffset = 0, yOffset = 0 }
+	E.db.unitframe.units.arena.customTexts.Luckyone_HP = {
+		attachTextTo = 'Health',
+		enable = true,
+		font = Private.Font,
+		fontOutline = Private.Outline,
+		justifyH = 'LEFT',
+		size = 12,
+		text_format = '[luckyone:health:percent] • [health:current:shortvalue]',
+		xOffset = 3,
+		yOffset = 0
+	}
+	E.db.unitframe.units.arena.customTexts.Luckyone_Name = {
+		attachTextTo = 'Health',
+		enable = true,
+		font = Private.Font,
+		fontOutline = Private.Outline,
+		justifyH = 'RIGHT',
+		size = 12,
+		text_format = '[luckyone:name:last-classcolor]',
+		xOffset = -3,
+		yOffset = 0
+	}
+	E.db.unitframe.units.arena.customTexts.Luckyone_Power = {
+		attachTextTo = 'Power',
+		enable = true,
+		font = Private.Font,
+		fontOutline = Private.Outline,
+		justifyH = 'CENTER',
+		size = 11,
+		text_format = '[luckyone:power:percent-color]',
+		xOffset = 0,
+		yOffset = 0
+	}
 
 	E.db.unitframe.units.arena.buffs.countFont = Private.Font
 	E.db.unitframe.units.arena.buffs.countFontSize = 10
@@ -741,9 +771,39 @@ function Private:Layout_TheWarWithin(layout)
 
 	-- Shared Boss
 	E.db.unitframe.units.boss.customTexts = {}
-	E.db.unitframe.units.boss.customTexts.Luckyone_HP = { attachTextTo = 'Health', enable = true, font = Private.Font, fontOutline = Private.Outline, justifyH = 'LEFT', size = 12, text_format = '[luckyone:health:percent] • [health:current:shortvalue]', xOffset = 3, yOffset = 0 }
-	E.db.unitframe.units.boss.customTexts.Luckyone_Name = { attachTextTo = 'Health', enable = true, font = Private.Font, fontOutline = Private.Outline, justifyH = 'RIGHT', size = 12, text_format = '[luckyone:name:last-classcolor]', xOffset = -3, yOffset = 0 }
-	E.db.unitframe.units.boss.customTexts.Luckyone_Power = { attachTextTo = 'Power', enable = true, font = Private.Font, fontOutline = Private.Outline, justifyH = 'CENTER', size = 11, text_format = '[luckyone:power:percent-color]', xOffset = 0, yOffset = 0 }
+	E.db.unitframe.units.boss.customTexts.Luckyone_HP = {
+		attachTextTo = 'Health',
+		enable = true,
+		font = Private.Font,
+		fontOutline = Private.Outline,
+		justifyH = 'LEFT',
+		size = 12,
+		text_format = '[luckyone:health:percent] • [health:current:shortvalue]',
+		xOffset = 3,
+		yOffset = 0
+	}
+	E.db.unitframe.units.boss.customTexts.Luckyone_Name = {
+		attachTextTo = 'Health',
+		enable = true,
+		font = Private.Font,
+		fontOutline = Private.Outline,
+		justifyH = 'RIGHT',
+		size = 12,
+		text_format = '[luckyone:name:last-classcolor]',
+		xOffset = -3,
+		yOffset = 0
+	}
+	E.db.unitframe.units.boss.customTexts.Luckyone_Power = {
+		attachTextTo = 'Power',
+		enable = true,
+		font = Private.Font,
+		fontOutline = Private.Outline,
+		justifyH = 'CENTER',
+		size = 11,
+		text_format = '[luckyone:power:percent-color]',
+		xOffset = 0,
+		yOffset = 0
+	}
 
 	E.db.unitframe.units.boss.buffIndicator.enable = false
 	E.db.unitframe.units.boss.buffs.countFont = Private.Font
@@ -812,8 +872,28 @@ function Private:Layout_TheWarWithin(layout)
 
 	-- Shared Focus
 	E.db.unitframe.units.focus.customTexts = {}
-	E.db.unitframe.units.focus.customTexts.Luckyone_Name = { attachTextTo = 'Frame', enable = true, font = Private.Font, fontOutline = Private.Outline, justifyH = 'CENTER', size = 12, text_format = '[luckyone:name:last-classcolor]', xOffset = 0, yOffset = 0 }
-	E.db.unitframe.units.focus.customTexts.Luckyone_Power = { attachTextTo = 'Power', enable = true, font = Private.Font, fontOutline = Private.Outline, justifyH = 'CENTER', size = 11, text_format = '[luckyone:power:percent-color]', xOffset = 0, yOffset = 0 }
+	E.db.unitframe.units.focus.customTexts.Luckyone_Name = {
+		attachTextTo = 'Frame',
+		enable = true,
+		font = Private.Font,
+		fontOutline = Private.Outline,
+		justifyH = 'CENTER',
+		size = 12,
+		text_format = '[luckyone:name:last-classcolor]',
+		xOffset = 0,
+		yOffset = 0
+	}
+	E.db.unitframe.units.focus.customTexts.Luckyone_Power = {
+		attachTextTo = 'Power',
+		enable = true,
+		font = Private.Font,
+		fontOutline = Private.Outline,
+		justifyH = 'CENTER',
+		size = 11,
+		text_format = '[luckyone:power:percent-color]',
+		xOffset = 0,
+		yOffset = 0
+	}
 
 	E.db.unitframe.units.focus.buffs.anchorPoint = 'TOPRIGHT'
 	E.db.unitframe.units.focus.buffs.attachTo = 'DEBUFFS'
@@ -877,8 +957,28 @@ function Private:Layout_TheWarWithin(layout)
 
 	-- Shared Pet
 	E.db.unitframe.units.pet.customTexts = {}
-	E.db.unitframe.units.pet.customTexts.Luckyone_Name = { attachTextTo = 'Frame', enable = true, font = Private.Font, fontOutline = Private.Outline, justifyH = 'CENTER', size = 12, text_format = '[luckyone:pet:name-and-happiness]', xOffset = 0, yOffset = 0 }
-	E.db.unitframe.units.pet.customTexts.Luckyone_Power = { attachTextTo = 'Power', enable = true, font = Private.Font, fontOutline = Private.Outline, justifyH = 'CENTER', size = 11, text_format = '[luckyone:power:percent-color]', xOffset = 0, yOffset = 0 }
+	E.db.unitframe.units.pet.customTexts.Luckyone_Name = {
+		attachTextTo = 'Frame',
+		enable = true,
+		font = Private.Font,
+		fontOutline = Private.Outline,
+		justifyH = 'CENTER',
+		size = 12,
+		text_format = '[luckyone:pet:name-and-happiness]',
+		xOffset = 0,
+		yOffset = 0
+	}
+	E.db.unitframe.units.pet.customTexts.Luckyone_Power = {
+		attachTextTo = 'Power',
+		enable = true,
+		font = Private.Font,
+		fontOutline = Private.Outline,
+		justifyH = 'CENTER',
+		size = 11,
+		text_format = '[luckyone:power:percent-color]',
+		xOffset = 0,
+		yOffset = 0
+	}
 
 	E.db.unitframe.units.pet.castbar.enable = false
 	E.db.unitframe.units.pet.disableMouseoverGlow = true
@@ -899,9 +999,39 @@ function Private:Layout_TheWarWithin(layout)
 
 	-- Shared Player
 	E.db.unitframe.units.player.customTexts = {}
-	E.db.unitframe.units.player.customTexts.Luckyone_HP = { attachTextTo = 'Frame', enable = true, font = Private.Font, fontOutline = Private.Outline, justifyH = 'RIGHT', size = 12, text_format = '[health:current:shortvalue] • [luckyone:health:percent]', xOffset = -2, yOffset = 0 }
-	E.db.unitframe.units.player.customTexts.Luckyone_Name = { attachTextTo = 'Frame', enable = true, font = Private.Font, fontOutline = Private.Outline, justifyH = 'LEFT', size = 12, text_format = '[luckyone:name:last-classcolor]', xOffset = 5, yOffset = 0 }
-	E.db.unitframe.units.player.customTexts.Luckyone_Power = { attachTextTo = 'Power', enable = true, font = Private.Font, fontOutline = Private.Outline, justifyH = 'CENTER', size = 11, text_format = '', xOffset = 0, yOffset = 0 }
+	E.db.unitframe.units.player.customTexts.Luckyone_HP = {
+		attachTextTo = 'Frame',
+		enable = true,
+		font = Private.Font,
+		fontOutline = Private.Outline,
+		justifyH = 'RIGHT',
+		size = 12,
+		text_format = '[health:current:shortvalue] • [luckyone:health:percent]',
+		xOffset = -2,
+		yOffset = 0
+	}
+	E.db.unitframe.units.player.customTexts.Luckyone_Name = {
+		attachTextTo = 'Frame',
+		enable = true,
+		font = Private.Font,
+		fontOutline = Private.Outline,
+		justifyH = 'LEFT',
+		size = 12,
+		text_format = '[luckyone:name:last-classcolor]',
+		xOffset = 5,
+		yOffset = 0
+	}
+	E.db.unitframe.units.player.customTexts.Luckyone_Power = {
+		attachTextTo = 'Power',
+		enable = true,
+		font = Private.Font,
+		fontOutline = Private.Outline,
+		justifyH = 'CENTER',
+		size = 11,
+		text_format = '',
+		xOffset = 0,
+		yOffset = 0
+	}
 
 	E.db.unitframe.units.player.aurabar.enable = false
 	E.db.unitframe.units.player.buffs.attachTo = 'FRAME'
@@ -972,9 +1102,39 @@ function Private:Layout_TheWarWithin(layout)
 
 	-- Shared Target
 	E.db.unitframe.units.target.customTexts = {}
-	E.db.unitframe.units.target.customTexts.Luckyone_HP = { attachTextTo = 'Frame', enable = true, font = Private.Font, fontOutline = Private.Outline, justifyH = 'LEFT', size = 12, text_format = '[luckyone:health:percent] • [health:current:shortvalue]', xOffset = 3, yOffset = 0 }
-	E.db.unitframe.units.target.customTexts.Luckyone_Name = { attachTextTo = 'Frame', enable = true, font = Private.Font, fontOutline = Private.Outline, justifyH = 'RIGHT', size = 12, text_format = (dev and '[luckyone:name:last-classcolor][ |r» >luckyone:target:last-classcolor]') or '[luckyone:name:last-classcolor]', xOffset = -3, yOffset = 0 }
-	E.db.unitframe.units.target.customTexts.Luckyone_Power = { attachTextTo = 'Power', enable = true, font = Private.Font, fontOutline = Private.Outline, justifyH = 'CENTER', size = 11, text_format = '[luckyone:power:percent-color]', xOffset = 0, yOffset = 0 }
+	E.db.unitframe.units.target.customTexts.Luckyone_HP = {
+		attachTextTo = 'Frame',
+		enable = true,
+		font = Private.Font,
+		fontOutline = Private.Outline,
+		justifyH = 'LEFT',
+		size = 12,
+		text_format = '[luckyone:health:percent] • [health:current:shortvalue]',
+		xOffset = 3,
+		yOffset = 0
+	}
+	E.db.unitframe.units.target.customTexts.Luckyone_Name = {
+		attachTextTo = 'Frame',
+		enable = true,
+		font = Private.Font,
+		fontOutline = Private.Outline,
+		justifyH = 'RIGHT',
+		size = 12,
+		text_format = (dev and '[luckyone:name:last-classcolor][ |r» >luckyone:target:last-classcolor]') or '[luckyone:name:last-classcolor]',
+		xOffset = -3,
+		yOffset = 0
+	}
+	E.db.unitframe.units.target.customTexts.Luckyone_Power = {
+		attachTextTo = 'Power',
+		enable = true,
+		font = Private.Font,
+		fontOutline = Private.Outline,
+		justifyH = 'CENTER',
+		size = 11,
+		text_format = '[luckyone:power:percent-color]',
+		xOffset = 0,
+		yOffset = 0
+	}
 
 	E.db.unitframe.units.target.aurabar.enable = false
 	E.db.unitframe.units.target.buffs.attachTo = 'DEBUFFS'
@@ -1042,7 +1202,17 @@ function Private:Layout_TheWarWithin(layout)
 
 	-- Shared TargetTarget
 	E.db.unitframe.units.targettarget.customTexts = {}
-	E.db.unitframe.units.targettarget.customTexts.Luckyone_Name = { attachTextTo = 'Frame', enable = true, font = Private.Font, fontOutline = Private.Outline, justifyH = 'CENTER', size = 12, text_format = '[luckyone:name:last-classcolor]', xOffset = 0, yOffset = 0 }
+	E.db.unitframe.units.targettarget.customTexts.Luckyone_Name = {
+		attachTextTo = 'Frame',
+		enable = true,
+		font = Private.Font,
+		fontOutline = Private.Outline,
+		justifyH = 'CENTER',
+		size = 12,
+		text_format = '[luckyone:name:last-classcolor]',
+		xOffset = 0,
+		yOffset = 0
+	}
 
 	E.db.unitframe.units.targettarget.enable = (not dev)
 	E.db.unitframe.units.targettarget.debuffs.enable = false
@@ -1121,7 +1291,17 @@ function Private:Layout_TheWarWithin(layout)
 
 	-- Shared Party
 	E.db.unitframe.units.party.customTexts = {}
-	E.db.unitframe.units.party.customTexts.Luckyone_Name = { attachTextTo = 'Health', enable = true, font = Private.Font, fontOutline = Private.Outline, justifyH = 'CENTER', size = 12, text_format = '', xOffset = 0, yOffset = 0 }
+	E.db.unitframe.units.party.customTexts.Luckyone_Name = {
+		attachTextTo = 'Health',
+		enable = true,
+		font = Private.Font,
+		fontOutline = Private.Outline,
+		justifyH = 'CENTER',
+		size = 12,
+		text_format = '',
+		xOffset = 0,
+		yOffset = 0
+	}
 
 	E.db.unitframe.units.party.classbar.enable = false
 	E.db.unitframe.units.party.debuffs.enable = false
@@ -1331,7 +1511,17 @@ function Private:Layout_TheWarWithin(layout)
 
 		-- Main Party (Classic Only)
 		if E.Classic then
-			E.db.unitframe.units.party.customTexts.Luckyone_Power = { attachTextTo = 'Power', enable = true, font = Private.Font, fontOutline = Private.Outline, justifyH = 'CENTER', size = 11, text_format = '[luckyone:power:percent-color<%]', xOffset = 88, yOffset = 20 }
+			E.db.unitframe.units.party.customTexts.Luckyone_Power = {
+				attachTextTo = 'Power',
+				enable = true,
+				font = Private.Font,
+				fontOutline = Private.Outline,
+				justifyH = 'CENTER',
+				size = 11,
+				text_format = '[luckyone:power:percent-color<%]',
+				xOffset = 88,
+				yOffset = 20
+			}
 			E.db.unitframe.units.party.power.autoHide = true
 			E.db.unitframe.units.party.power.enable = true
 			E.db.unitframe.units.party.power.height = 4